Context - Implement Dynamic setting of ThemeContext value
We add a new dropdown to SearchParams with choices of colors that updates the ThemeContext value (since we're just tracking a simple string on it, its just the one value).  And we can see that that updates the button background color for the Search Params (front page) and Details pages.

It was also necessary to update Pet.js to use Link and not a standard a tag in order to avoid it unmounting/completely reloading the app and resetting state (including theme color.)